Field OrderID datatype change

+++ src/com/java2007/zhangjinnan/restaurant/dao/OrderDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface OrderDao {
     int save(Order order) throws SQLException;
 
-    int updateOrderStatus(int orderId) throws SQLException;
+    int updateOrderStatus(Long orderId) throws SQLException;
 
     List<Order> findAll() throws SQLException;
